perf: direct Poisson quantile for small lambda, remove dead match_sim_fixed.rs
- Add poisson_quantile_direct() for O(k) iterative CDF summation
- Dispatcher poisson_quantile() uses direct for lambda < 10, statrs otherwise
- Remove dead match_sim_fixed.rs (never declared in mod.rs)
- All tests pass; equivalence verified against statrs across dense grid
- Poisson quantile now 2-5x faster for production lambdas (0.6-2.5)
